improve sync status output format; reduce footprint of build tagged files in the agentsocket package
