+++ agent/agentsocket/service.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 package agentsocket
 
 import (
