+++ cli/testdata/TestSyncCommands_Golden/status_completed.golden
@@ -1,3 +1,6 @@
 Unit: test-unit
 Status: completed
 Ready: true
+
+Dependencies:
+No dependencies found
