+++ cli/testdata/TestSyncCommands_Golden/status_pending.golden
@@ -2,5 +2,6 @@ Unit: test-unit
 Status: pending
 Ready: false
 
+Dependencies:
 DEPENDS ON  REQUIRED STATUS  CURRENT STATUS  SATISFIED  
-dep-unit    completed        not registered  false
+dep-unit    completed        not registered  false
