+++ cli/testdata/TestSyncCommands_Golden/status_started.golden
@@ -1,3 +1,6 @@
 Unit: test-unit
 Status: started
 Ready: true
+
+Dependencies:
+No dependencies found
