+++ cli/testdata/TestSyncCommands_Golden/status_with_dependencies.golden
@@ -2,6 +2,7 @@ Unit: test-unit
 Status: pending
 Ready: false
 
+Dependencies:
 DEPENDS ON  REQUIRED STATUS  CURRENT STATUS  SATISFIED  
 dep-1       completed        completed       true       
-dep-2       completed        not registered  false
+dep-2       completed        not registered  false
